requirements.txt
Flask
rq
redis
opencv-python-headless
Pillow
numpy
requests
python-dotenv
openai
gunicorn
honcho
google-auth
google-auth-oauthlib 
google-auth-httplib2 
google-api-python-client